Events

+++ src/main/java/com/planification/wf/repository/UserRepository.java
@@ -1,5 +1,6 @@
 package com.planification.wf.repository;
 
+import com.planification.wf.entity.Events;
 import com.planification.wf.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -7,6 +8,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
+
+
 }
